Build in a clean build dir instead of removing build files afterwards
Removing files from the release tarball is not enough because
ansible-galaxy will add a FILES.json to the tarball that lists all files
in the collection that were present during build time. If any of these
files is not available during install, the ansible-galaxy will fail with
an error message, e.g. "ERROR! Unable to extract 'build/' from collection"