Upgraded OpenStackClient from PyPI on older operating systems
On Debian 10 (Buster) and Ubuntu 18.04 LTS (Bionic Beaver) role setup will
install OpenStack SDK 0.45.0, which got released with OpenStack Ussuri,
because version 0.45.0 is the last version that was published to PyPI for
Python 2. Newer OpenStack SDK releases like 0.45.0 require an upgrade of
OpenStackClient to work properly, thus OpenStackClient 5.2.1 will be
installed from PyPI as well. This will pull in an update for Python
package PyYAML which conflicts with the installed debian package
python(3)-yaml so a workaround has been implemented.